Add unix domain socket support for the RPCClient.
To not break backwards compatibility this commit makes host and port
optional and None by default and adds unix_socket_path argument.

Now it is possible to provide either host and port or path to a unix
domain socket. If one of those arguments are missing, AssertionError
will be raised.