Allow specification of logging path, and don't use the anonymous logging calls.
This is a minor change that allows the logging path to be overriden, and no longer uses the `logging.<level>()` calls. This was primarily done because calling the `logging.<level>()` calls can cause the logging library to sometimes install the `logging.basicConfig()` log handler, which leads to duplicate logging if another log-handler is installed.