Fix defaults for msgpack-python with respect to blobs
msgpack-python defaults to using binary strings, unless an encoding is
specified. In that case, `use_bin_type=True` needs to also be specified
to allow blobs to be serialized as well as strings.

mprpc specifies an encoding, but not `use_bin_type`, which this commit
rectifies. The default setting now allows Unicode strings well as
arbitrary blobs.