Fix Py2k method name type
The fix introduced in 227231f3 introduced another bug: By enabling
binary strings, Python 2 would serialize the method name as a binary
string, not UTF-8 as according to the standard.
An obvious way to make this fail is by running the server with Python 3
and the client with Python 2.

This conditionally converts strings to UTF-8 when using Python 2.

Signed-off-by: Martin Braun <martin.braun@ettus.com>